+++ src/group.rs
@@ -314,7 +314,9 @@ mod tests {
     #[cfg(feature = "async")]
     #[tokio::test]
     async fn group_metadata_write_read_async() {
-        let store = std::sync::Arc::new(crate::storage::store::AsyncMemoryStore::new());
+        let store = std::sync::Arc::new(crate::storage::store::AsyncObjectStore::new(
+            object_store::memory::InMemory::new(),
+        ));
         let group_path = "/group";
         let group = GroupBuilder::new()
             .build(store.clone(), group_path)